refactor(useWish): 위시 상태, 위시 이벤트 핸들러 모듈화
