+++ src/components/features/Goods/Detail/OptionSection.tsx
@@ -9,6 +9,7 @@ import {
 } from '@/api/hooks/useGetProductDetail';
 import { useGetProductOptions } from '@/api/hooks/useGetProductOptions';
 import { Button } from '@/components/common/Button';
+import { useWish } from '@/hooks/useWish';
 import { useAuth } from '@/provider/Auth';
 import { getDynamicPath, RouterPath } from '@/routes/path';
 import { orderHistorySessionStorage } from '@/utils/storage';
@@ -21,28 +22,16 @@ export const OptionSection = ({ productId }: Props) => {
   const { data: detail } = useGetProductDetail({ productId });
   const { data: options } = useGetProductOptions({ productId });
 
-  const [isWish, setIsWish] = useState(false);
   const [countAsString, setCountAsString] = useState('1');
   const totalPrice = useMemo(() => {
     return detail.price * Number(countAsString);
   }, [detail, countAsString]);
 
   const navigate = useNavigate();
+
   const authInfo = useAuth();
 
-  const handleWishClick = () => {
-    if (!isWish) {
-      // 위시 등록 api 요청
-
-      alert('위시에 담았어요!');
-    } else {
-      // 위시 삭제 api 요청
-
-      alert('취소! 위시에서 삭제할게요.');
-    }
-
-    setIsWish(!isWish);
-  };
+  const { isWish, handleWishClick } = useWish();
 
   const handleOrderClick = () => {
     if (!authInfo) {
