+++ src/components/features/Wish/WishItem.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
-import { useState } from 'react';
 import { AiFillHeart, AiOutlineHeart } from 'react-icons/ai';
 
 import { Image } from '@/components/common/Image';
+import { useWish } from '@/hooks/useWish';
 
 type Props = {
   name: string;
@@ -11,22 +11,10 @@ type Props = {
 };
 
 export const WishItem = ({ name, imageURL, price }: Props) => {
-  const [isWish, setIsWish] = useState(false);
+  const { isWish, handleWishClick } = useWish();
 
   const handleWishItemClick = () => {
     // 상품 상세 페이지로 이동
-  };
-
-  const handleWishClick = () => {
-    if (!isWish) {
-      // 위시 등록 api 요청
-      alert('위시에 담았어요!');
-    } else {
-      // 위시 삭제 api 요청
-      alert('취소! 위시에서 삭제할게요.');
-    }
-
-    setIsWish(!isWish);
   };
 
   return (
